+++ docs/changelog.rst
@@ -24,6 +24,10 @@ Bug fix release.
 - Network: Fix a race condition where two threads could try to free the same
   data simultaneously. (Fixes: :issue:`781`)
 
+- Backend API: Update :meth:`mopidy.backend.LibraryProvider.browse` signature
+  and docs to match how the core use the backend's browse method. (Fixes:
+  :issue:`833`)
+
 
 v0.19.3 (2014-08-03)
 ====================